Restore click and keyboard input
New visual keymap - change the way changes are conveyed to the keys

Because of reactive model, we need something simpler, it probably causes
more updates that I would like, but it's good enough for now. Basically
set a variable of the index for the last key that gets changed. Check
that variable in each of the keys. If the index matches then update the
displayname. Otherwise ignore the update.

Not the most efficient way to do it, but it fits the reactive model
until I can come up with something better.